Add background tileset
I added a background tileset using the provided sprites and created in
the program Tiled. This should be extensible to allow for different
backgrounds on a per level basis, and Tiled makes them fairly easy to
generate. My next plan is to add a config for each level that will
detail what tileset is to be used, as well as the starting positions
for the player, any enemies, any obstacles, and any oilslicks.