Adds individual level control in json file
The main thing this adds is for a json file to control all the aspects
of each level, so that they don't have to be hardcoded into the program.
This makes it much easier to come up with new level layouts and designs.

Also included here are several smaller tweaks, including:
 - Water terrains that slows a player down
 - Added sound effects for crossing water or breaking obstacles
 - Improve enemy firing mechanics so that they won't randomly unleash a
   barrage of bullets, but fire more sporadically
 - Remove level selection, since there can now be more than just 3
   levels. May be able to add in a level selector again in the future at
   some point.